size_t for count

+++ src/command.h
@@ -3,6 +3,7 @@
 #include <time.h>
 #include "sort.h"
 #include <chrono>
+#include <cstddef> //library for size_t
 using namespace std;
 using std::chrono::duration_cast;
 using std::chrono::duration;
@@ -17,5 +18,5 @@ void GenerateData(int a[], int n, int dataType);
 int isNum(char* str);
 int* readFile(char* file_name, int& n);
 void writeFile(char* file_name, int* array, int n);
-unsigned int count_comparisons(int* array, int n, string algro);
+size_t count_comparisons(int* array, int n, string algro);
 double runtime(int* array, int n, string algro);
